+++ src/main/java/fr/univbrest/dosi/spi/controller/QuestionEvaluationController.java
@@ -3,8 +3,11 @@ package fr.univbrest.dosi.spi.controller;
 import java.util.ArrayList;
 import java.util.List;
 
+import fr.univbrest.dosi.spi.exception.SPIException;
+import fr.univbrest.dosi.spi.exception.SpiExceptionCode;
 import fr.univbrest.dosi.spi.service.EvaluationService;
 import fr.univbrest.dosi.spi.service.RubriqueService;
+import fr.univbrest.dosi.spi.util.Connection;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -20,6 +23,8 @@ import fr.univbrest.dosi.spi.bean.Rubrique;
 import fr.univbrest.dosi.spi.bean.RubriqueEvaluation;
 import fr.univbrest.dosi.spi.service.QuestionEvaluationService;
 import fr.univbrest.dosi.spi.service.RubriqueEvaluationService;
+
+import javax.servlet.http.HttpServletRequest;
 
 @RestController
 @RequestMapping("/questionEvaluation")
@@ -38,7 +43,9 @@ public class QuestionEvaluationController {
     }
 
     @RequestMapping(method = RequestMethod.POST)
-    public boolean addQuestionEvaluation(@RequestBody final QuestionEvaluation questionEvaluation) {
+    public boolean addQuestionEvaluation(@RequestBody final QuestionEvaluation questionEvaluation,HttpServletRequest request) {
+        if(Connection.currentUser(request) == null || !Connection.currentUser(request).getRole().equals("Prof"))
+            throw  new SPIException(SpiExceptionCode.NOT_ENOUGH_RIGHT,"access denied");
         try {
             questionEvaluationService.addQuestionEvaluation(questionEvaluation);
             return true;
@@ -48,7 +55,9 @@ public class QuestionEvaluationController {
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-    public final boolean deleteQuestionEvaluation(@PathVariable(value = "id") final long id) {
+    public final boolean deleteQuestionEvaluation(@PathVariable(value = "id") final long id,HttpServletRequest request) {
+        if(Connection.currentUser(request) == null || !Connection.currentUser(request).getRole().equals("Prof"))
+            throw  new SPIException(SpiExceptionCode.NOT_ENOUGH_RIGHT,"access denied");
         try {
             questionEvaluationService.deleteQuestionEvaluation(id);
             return true;
@@ -58,25 +67,33 @@ public class QuestionEvaluationController {
     }
 
     @RequestMapping(value = "notAffectedQuestion/evaluation/{idEvaluation}/rubrique/{idRubrique}")
-    public List<Question> getQuestionsNotAffected(@PathVariable(value = "idEvaluation") int idEvaluation, @PathVariable(value = "idRubrique") int idRubrique) {
+    public List<Question> getQuestionsNotAffected(@PathVariable(value = "idEvaluation") int idEvaluation, @PathVariable(value = "idRubrique") int idRubrique,HttpServletRequest request) {
+        if(Connection.currentUser(request) == null || !Connection.currentUser(request).getRole().equals("Prof"))
+            throw  new SPIException(SpiExceptionCode.NOT_ENOUGH_RIGHT,"access denied");
         return questionEvaluationService.getQuestionsNotAffectedByRubriqueEvaluation(rubriqueEvaluationService.findByEvaluationAndRubrique(evaluationService.getEvaluation(idEvaluation), rubriqueService.getRubrique(idRubrique)).get(0));
     }
 
 
     @RequestMapping(method = RequestMethod.GET)
-    public List<QuestionEvaluation> getAllQuestionsEvaluations() {
+    public List<QuestionEvaluation> getAllQuestionsEvaluations(HttpServletRequest request) {
+        if(Connection.currentUser(request) == null || !Connection.currentUser(request).getRole().equals("Prof"))
+            throw  new SPIException(SpiExceptionCode.NOT_ENOUGH_RIGHT,"access denied");
         return questionEvaluationService.getAllQuestionsEvaluations();
     }
 
 
     @RequestMapping(value = "/{idquestionEvaluation}")
-    public final QuestionEvaluation getQuestionEvaluation(@PathVariable(value = "idquestionEvaluation") final long idQuestionEvaluation) {
+    public final QuestionEvaluation getQuestionEvaluation(@PathVariable(value = "idquestionEvaluation") final long idQuestionEvaluation,HttpServletRequest request) {
+        if(Connection.currentUser(request) == null || !Connection.currentUser(request).getRole().equals("Prof"))
+            throw  new SPIException(SpiExceptionCode.NOT_ENOUGH_RIGHT,"access denied");
         return questionEvaluationService.getQuestionEvaluation(idQuestionEvaluation);
     }
 
 
     @RequestMapping(value = "affectedQuestion/evaluation/{idEvaluation}/rubrique/{idRubrique}")
-    public List<QuestionEvaluation> getQuestionsAffected(@PathVariable(value = "idEvaluation") int idEvaluation, @PathVariable(value = "idRubrique") int idRubrique) {
+    public List<QuestionEvaluation> getQuestionsAffected(@PathVariable(value = "idEvaluation") int idEvaluation, @PathVariable(value = "idRubrique") int idRubrique,HttpServletRequest request) {
+        if(Connection.currentUser(request) == null || !Connection.currentUser(request).getRole().equals("Prof"))
+            throw  new SPIException(SpiExceptionCode.NOT_ENOUGH_RIGHT,"access denied");
 
         return  questionEvaluationService.getQuestionEvaluationByRubriqueEvaluation(rubriqueEvaluationService.findByEvaluationAndRubrique(evaluationService.getEvaluation(idEvaluation), rubriqueService.getRubrique(idRubrique)).get(0));
         /*List<Question> questionList = new ArrayList<Question>();
